refactor: rename header component
